[TAN-2514] Changed name of attribute and merged two test files
