+++ back/engines/commercial/bulk_import_ideas/app/services/bulk_import_ideas/patches/side_fx_idea_service.rb
@@ -7,7 +7,7 @@ module BulkImportIdeas
         return unless idea.author && idea.idea_import&.user_created
 
         # Log a user 'create' activity - when the idea has been published
-        LogActivityJob.perform_later(idea.author, 'created', current_user, idea.author.created_at.to_i, payload: { type: 'imported' })
+        LogActivityJob.perform_later(idea.author, 'created', current_user, idea.author.created_at.to_i, payload: { flow: 'importer' })
       end
 
       private
