+++ back/engines/commercial/bulk_import_ideas/spec/services/patches/side_fx_idea_service_spec.rb
@@ -20,7 +20,7 @@ describe SideFxIdeaService do
                 'created',
                 admin_user,
                 author.created_at.to_i,
-                payload: { type: 'imported' }
+                payload: { flow: 'importer' }
               )
               .exactly(1).times
     end
